ECV-552 Move Summary Headers to CSV exporter
